% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{birth_weight_fasting_glucose_default}
\alias{birth_weight_fasting_glucose_default}
\title{A dataset containing results for reproducing Figure 17 in Section 6 of the SMMR article.}
\format{
Matrix containing posterior samples of the causal effect from exposure
to outcome for the birth weight versus adult fasting glucose example. The
inference is performed using the default BayesMR settings.
}
\usage{
birth_weight_fasting_glucose_default
}
\description{
A dataset containing results for reproducing Figure 17 in Section 6 of the SMMR article.
}
\keyword{datasets}
