% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{birth_weight_fasting_glucose_IV}
\alias{birth_weight_fasting_glucose_IV}
\title{A dataset containing results for reproducing Figure 18 in Section 6 of the SMMR article.}
\format{
Matrix containing posterior samples of the causal effect from exposure
to outcome for the birth weight versus adult fasting glucose example. The
inference is performed by incorporating the strict IV assumptions into the
BayesMR model (set model == 1 in configuration file).
}
\usage{
birth_weight_fasting_glucose_IV
}
\description{
A dataset containing results for reproducing Figure 18 in Section 6 of the SMMR article.
}
\keyword{datasets}
