% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coffee_consumption_smoking_direction_posterior}
\alias{coffee_consumption_smoking_direction_posterior}
\title{A dataset containing results for reproducing Figure 22a in Section 6 of the SMMR article.}
\format{
Data frame containing posterior samples for the estimated causal effect
of coffee consumption on the heaviness of smoking.
}
\usage{
coffee_consumption_smoking_direction_posterior
}
\description{
A dataset containing results for reproducing Figure 22a in Section 6 of the SMMR article.
}
\keyword{datasets}
