% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolyChord_interface.R
\name{derive_polychord_Bayes_factor}
\alias{derive_polychord_Bayes_factor}
\title{Function for comparing the log-evidence from two BayesMR runs corresponding
to both directions of the causal link and deriving the Bayes factor.}
\usage{
derive_polychord_Bayes_factor(filename_expected, filename_reverse)
}
\arguments{
\item{filename_expected}{Name of PolyChord output file containing log-evidence
for the BayesMR run in the expected causal link direction.}

\item{filename_reverse}{Name of PolyChord output file containing log-evidence
for the BayesMR run in the reverse causal link direction.}
}
\value{
List containing Bayes factor together with its lower and upper bar,
as well as the probability of the expected and reverse direction.
}
\description{
Function for comparing the log-evidence from two BayesMR runs corresponding
to both directions of the causal link and deriving the Bayes factor.
}
