% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{estimation_robustness}
\alias{estimation_robustness}
\title{A dataset containing results for reproducing Figure 15 in Section 5 of the SMMR article.}
\format{
A data frame with six columns containing posterior PolyChord beta
(causal effect) samples for six different scenarios in terms of IV validity
(100\%, 80\%, 60\%, 40\%, 20\%, 0\%) to show the estimation robustness.
}
\usage{
estimation_robustness
}
\description{
A dataset containing results for reproducing Figure 15 in Section 5 of the SMMR article.
}
\keyword{datasets}
