% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolyChord_interface.R
\name{extract_beta_PolyChord_samples}
\alias{extract_beta_PolyChord_samples}
\title{Function for extracting the causal effect from exposure to outcome from a
data frame of PolyChord samples.}
\usage{
extract_beta_PolyChord_samples(
  data,
  idx_sbeta = ncol(data) - 4,
  idx_sigma_X = ncol(data) - 1,
  idx_sigma_Y = ncol(data),
  slab_precision = 1,
  spike_precision = 100
)
}
\arguments{
\item{data}{Data frame containing PolyChord samples, typically read using read.table.}

\item{idx_sbeta}{Integer index of sbeta variable.}

\item{idx_sigma_X}{Integer index of sigma_X variable.}

\item{idx_sigma_Y}{Integer index of sigma_Y variable.}

\item{slab_precision}{Numeric precision of slab component.}

\item{spike_precision}{Numeric precision of spike component.}
}
\value{
Numeric vector containing samples of beta (the causal effect from
exposure to outcome) estimates in the original scale.
}
\description{
Function for extracting the causal effect from exposure to outcome from a
data frame of PolyChord samples.
}
