% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_data_BayesMR_model}
\alias{generate_data_BayesMR_model}
\title{Function to generate data from a BayesMR generating model.}
\usage{
generate_data_BayesMR_model(
  N,
  theta,
  gamma,
  alpha,
  beta,
  kappa_X,
  kappa_Y,
  sigma_X,
  sigma_Y,
  seed = NULL,
  n = 2
)
}
\arguments{
\item{N}{Integer number of observations.}

\item{theta}{Numeric expected allele frequencies for the genetic variables.}

\item{gamma}{Numeric direct causal effect from genetic variants to exposure.}

\item{alpha}{Numeric direct causal effect from genetic variants to outcome.}

\item{beta}{Numeric causal effect from exposure to outcome.}

\item{kappa_X}{Numeric confounding effect from confounder to exposure.}

\item{kappa_Y}{Numeric confounding effect from confounder to outcome.}

\item{sigma_X}{Numeric intrinsic standard deviation of exposure.}

\item{sigma_Y}{Numeric intrinsic standard deviation of outcome.}

\item{seed}{Integer random seed for reproducibility.}

\item{n}{Integer number of alleles (trials) for the binomial genetic variables.}
}
\value{

}
\description{
Function to generate data from a BayesMR generating model.
}
\examples{
generate_data_BayesMR_model(N = 1000, theta = c(0.5, 0.3), gamma = c(1, 1), 
alpha = c(0, 0.05), beta = 1, kappa_X = 1, kappa_Y = 1, sigma_X = 1, sigma_Y = 1)
}
