% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model_averaging_robustness}
\alias{model_averaging_robustness}
\title{A dataset containing results for reproducing Table 1 in Section 5 of the SMMR article.}
\format{
A data frame containin the information of Table 1 which shows how
robust the model averaging performed by BayesMR to pleiotropy.
\describe{
\item{delta}{The size and direction of the pleiotropic effects.}
\item{expected_gamma_hat}{Genetic association with exposure in expected direction.}
\item{expected_gamma_std_err}{Standard error for genetic association with exposure in expected direction.}
\item{expected_Gamma_hat}{Genetic association with outcome in expected direction.}
\item{expected_Gamma_std_err}{Standard error for genetic association with outcome in expected direction.}
\item{reverse_gamma_hat}{Genetic association with exposure in reverse direction.}
\item{reverse_gamma_std_err}{Standard error for genetic association with exposure in reverse direction.}
\item{reverse_Gamma_hat}{Genetic association with outcome in reverse direction.}
\item{reverse_Gamma_std_err}{Standard error for genetic association with outcome in reverse direction.}
\item{probability_expected}{BayesMR probability of expected as opposed to reverse direction.}
}
}
\usage{
model_averaging_robustness
}
\description{
A dataset containing results for reproducing Table 1 in Section 5 of the SMMR article.
}
\keyword{datasets}
