% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_sufficient_statistics.R
\name{MR_regression_coefficients_to_moments}
\alias{MR_regression_coefficients_to_moments}
\title{Function to extract first-order and second-order estimated moments from
summary data.}
\usage{
MR_regression_coefficients_to_moments(
  J,
  beta_XG,
  sigma_XG,
  obs_XG,
  beta_YG,
  sigma_YG,
  obs_YG,
  beta_YX,
  obs_YX,
  EAF = rep(0.5, J),
  n = 2
)
}
\arguments{
\item{J}{Integer number of instrumental variables.}

\item{beta_XG}{Numeric vector of estimated G -> X regression terms.}

\item{sigma_XG}{Numeric vector of estimated G -> X standard errors.}

\item{obs_XG}{Integer sample size in G -> X regression (GWA) study.}

\item{beta_YG}{Numeric vector of estimated G -> Y regression terms.}

\item{sigma_YG}{Numeric vector of estimated G -> Y standard errors.}

\item{obs_YG}{Integer sample size in G -> Y regression (GWA) study.}

\item{beta_YX}{Numeric vector of estimated X -> Y regression value.}

\item{obs_YX}{Integer sample size in X -> Y regression (observational) study.}

\item{EAF}{Numeric vector of effect allele frequencies}

\item{n}{Integer number of alleles (trials) for binomial genetic variant.}
}
\value{
Numeric matrix of size (J+3)x(J+3) containing the first-order and
second-order moments of the (J+2) vector (G, X, Y).
}
\description{
Function to extract first-order and second-order estimated moments from
summary data.
}
\examples{
MR_regression_coefficients_to_moments(1, 1, 1e-3, 1000, 1, 1e-3, 1000, 1, 1e-3)
}
