% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{near_LCD_example}
\alias{near_LCD_example}
\title{A dataset containing results for reproducing Figure 14 in Section 4 of the SMMR article.}
\format{
A three-level list containing posterior PolyChord samples in the
expected and reverse directions for the near LCD example as well as the
log-evidence comparison between the two directions.
}
\usage{
near_LCD_example
}
\description{
A dataset containing results for reproducing Figure 14 in Section 4 of the SMMR article.
}
\keyword{datasets}
