% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Parkinson_BMI_posterior}
\alias{Parkinson_BMI_posterior}
\title{A dataset containing results for reproducing Figures 19-21 in Section 6 of the SMMR article.}
\format{
Matrix containing posterior samples of the causal effect from exposure
to outcome for the risk of Parkinson's disease versus BMI example.
}
\usage{
Parkinson_BMI_posterior
}
\description{
A dataset containing results for reproducing Figures 19-21 in Section 6 of the SMMR article.
}
\keyword{datasets}
