% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quantile_spike_and_slab_vectorized}
\alias{quantile_spike_and_slab_vectorized}
\title{Vectorized version of function for computing spike-and-slab prior quantiles.}
\usage{
quantile_spike_and_slab_vectorized(p_vec, x_vec, slab_vec = 1, spike_vec = 100)
}
\arguments{
\item{p_vec}{Numeric vector of probability values between 0 and 1.}

\item{x_vec}{Numeric vector of weights for spike-and-slab mixtures, all
between 0 (spike) and 1 (slab). Alternatively, a single value to be used for
all p in p_vec.}

\item{slab_vec}{Numeric vector of slab component precisions. Alternatively,
a single value to be used for all p in p_vec.}

\item{spike_vec}{Numeric vector of spike component precisions. Alternatively,
a single value to be used for all p in p_vec.}
}
\value{
Vector of spike-and-slab quantile for given probability values.
}
\description{
Vectorized version of function for computing spike-and-slab prior quantiles.
}
\examples{
quantile_spike_and_slab_vectorized(c(0.9, 0.5), c(0.5, 0.3))
}
