% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolyChord_interface.R
\name{read_PolyChord_log_evidence}
\alias{read_PolyChord_log_evidence}
\title{Function for reading the log-evidence from PolyChord output.}
\usage{
read_PolyChord_log_evidence(filename)
}
\arguments{
\item{filename}{Name of the PolyChord file}
}
\value{
List containing estimated log-evidence, standard error, as well
as derived lower and upper 95\% CI error bars.
}
\description{
Function for reading the log-evidence from PolyChord output.
}
