% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolyChord_interface.R
\name{read_PolyChord_samples}
\alias{read_PolyChord_samples}
\title{Read data.frame from a PolyChord output file and name the columns appropriately.}
\usage{
read_PolyChord_samples(
  filename,
  slab_precision = 1,
  spike_precision = 100,
  quantile_transform = TRUE
)
}
\arguments{
\item{filename}{Should be of the form {root}.txt or {root}_equal_weights.txt (see PolyChord documentation)}

\item{slab_precision}{Numeric precision of slab component in the BayesMR priors.}

\item{spike_precision}{Numeric precision of spike component in the BayesMR priors.}

\item{quantile_transform}{Logical flag (true by default) indicating whether
the samples read from the PolyChord output should be transformed back from the
uniform hypercube to the parameters' distributions via the quantile function.}
}
\value{
A data frame containing the PolyChord samples for the BayesMR model variables.
}
\description{
Read data.frame from a PolyChord output file and name the columns appropriately.
}
