% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sensitivity_to_hyperparameters}
\alias{sensitivity_to_hyperparameters}
\title{A dataset containing results for reproducing Figure 16 in Section 5 of the SMMR article.}
\format{
A data frame containing two columns for showing the sensitivity of
the estimation wrt hyperparameters, in this case the spike precision.
\describe{
\item{spike}{Exponent of spike component precision (10^exponent).}
\item{beta}{Posterior PolyChord beta (causal effect) samples.}
}
}
\usage{
sensitivity_to_hyperparameters
}
\description{
A dataset containing results for reproducing Figure 16 in Section 5 of the SMMR article.
}
\keyword{datasets}
