% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sensitivity_to_nonlinearity}
\alias{sensitivity_to_nonlinearity}
\title{A dataset containing results for reproducing Table 2 in Section 5 of the SMMR article.}
\format{
A three level list which contains PolyChord posterior samples and
estimated log-evidences in both the expected and reverse direction for seven
different values of the nonlinearity parameter A.
}
\usage{
sensitivity_to_nonlinearity
}
\description{
A dataset containing results for reproducing Table 2 in Section 5 of the SMMR article.
}
\keyword{datasets}
