% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sensitivity_to_nonnormality}
\alias{sensitivity_to_nonnormality}
\title{A dataset containing results for reproducing Table 3 in Section 5 of the SMMR article.}
\format{
A three level list which contains PolyChord posterior samples and
estimated log-evidences in both the expected and reverse direction for seven
different values of the nonnormality parameter nu.
}
\usage{
sensitivity_to_nonnormality
}
\description{
A dataset containing results for reproducing Table 3 in Section 5 of the SMMR article.
}
\keyword{datasets}
