% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smoking_coffee_consumption_direction_posterior}
\alias{smoking_coffee_consumption_direction_posterior}
\title{A dataset containing results for reproducing Figure 22b in Section 6 of the SMMR article.}
\format{
Data frame containing posterior samples for the estimated causal effect
of smoking on coffee consumption.
}
\usage{
smoking_coffee_consumption_direction_posterior
}
\description{
A dataset containing results for reproducing Figure 22b in Section 6 of the SMMR article.
}
\keyword{datasets}
