% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_BayesMR_configuration_file.R
\name{write_BayesMR_configuration_file}
\alias{write_BayesMR_configuration_file}
\title{Function for automatically writing BayesMR configuration files for a given problem.}
\usage{
write_BayesMR_configuration_file(
  config_filename,
  SS_filename,
  sigma_G_filename,
  num_instruments,
  num_observations,
  slab_precision = 1,
  spike_precision = 100,
  model_type = 0,
  PolyChord_control = NULL
)
}
\arguments{
\item{config_filename}{String indicating the name of the configuration file.}

\item{SS_filename}{String indicating the file where the sufficient statistics are saved.}

\item{sigma_G_filename}{String indicating the file where the instrument frequency statistics are saved.}

\item{num_instruments}{Integer number of instruments for the given problem.}

\item{num_observations}{Integer number of observations in the provided sample.}

\item{slab_precision}{Numeric precision of prior slab component.}

\item{spike_precision}{Numeric precision of prior spike component.}

\item{model_type}{Integer indicating the type of prior considered (0 = spike-and-slab, 1 = Gaussian).}

\item{PolyChord_control}{List of PolyChord control parameters (see details section).}
}
\description{
Function for automatically writing BayesMR configuration files for a given problem.
}
\details{
PolyChord control parameters
}
\examples{
write_BayesMR_configuration_file(
  config_filename = "ini/BayesMR.ini",
  SS_filename = "inst/extdata/BayesMR_SS.txt",
  sigma_G_filename = "inst/extdata/BayesMR_sigma_G.txt",
  num_instruments = 1, num_observations = 1000
)
}
